+++ src/xpn_client/nfi/nfi_mpi_server/nfi_mpi_server_comm.cpp
@@ -303,7 +303,7 @@ int64_t nfi_mpi_server_comm::write_data(const void *data, int64_t size) {
     return size;
 }
 
-int64_t nfi_mpi_server_comm::read_data(void *data, ssize_t size) {
+int64_t nfi_mpi_server_comm::read_data(void *data, int64_t size) {
     XPN_PROFILE_FUNCTION_ARGS(size);
     int ret;
     MPI_Status status;
